requirements.txt
# Core web framework
flask>=2.3.0
flask-cors>=4.0.0

# AI/ML APIs and models
cohere>=4.0.0
sentence-transformers>=2.2.0

# Vector database
chromadb>=0.4.0

# Testing framework
pytest>=7.0.0

# HTTP requests
requests>=2.28.0

# Environment and configuration
python-dotenv>=1.0.0